requirements.txt
pandas
google-searchconsole
xlsxwriter